// src/app/types/types.ts
export interface SchemaSeo {
  cityDeparture: any;
  cityArrival: any;
  routesUrl: string;
  routesStopsCity: {
      cityDeparture: any;
      cityArrival: any;
  };
  routesStopsStation: {
      cityDeparture: any;
      cityArrival: any;
  };
  routesLength: number;
  routesPrice: number;
  ratingValue: any;
  reviewCount: any;
}